+++ pentobi/AndroidUtils.cpp
@@ -114,6 +114,16 @@ float AndroidUtils::getDensity()
 }
 #endif
 
+void AndroidUtils::quit()
+{
+#ifdef Q_OS_ANDROID
+    QAndroidJniObject::callStaticMethod<void>(
+                "java/lang/System", "exit", "(I)V", 0);
+#else
+    Qt::quit();
+#endif
+}
+
 void AndroidUtils::scanFile([[maybe_unused]] const QString& pathname)
 {
 #ifdef Q_OS_ANDROID